dpfj

+++ drawkorchart.py
@@ -51,7 +51,7 @@ def income_chart(input_ticker, income_df, income_df_q):
             for y_data, color in zip(y_data_line, marker_colors): 
                 fig.add_trace(go.Scatter(mode='lines+markers+text', 
                                             name = y_data, x =  income_df.index, y= income_df.loc[:,y_data],
-                                            text= int(income_df[y_data]), textposition = 'top center', marker_color = color),
+                                            text= income_df[y_data].astype(int), textposition = 'top center', marker_color = color),
                                             secondary_y = True)
             fig.update_traces(texttemplate='%{text:.3s}') 
             fig.update_yaxes(title_text='Profit', range=[0, max(income_df.loc[:,y_data_bar[0]])*2], secondary_y = False)
